feature-review-model
updated codes in all models

+++ backend/App/controllers/shops-controller.js
@@ -14,28 +14,14 @@ shopsCltr.register = async (req, res) => {
     console.log(req.user)
     const shop = new Shop(body)
     const response = await shop.save()
-    res.status(201).json(shop)
+    res.status(201).json(response)
 } catch (err) {
     console.error(err)
     res.status(500).json({ error: 'Internal server error' })
 }
 }
 
-shopsCltr.update = async (req, res) => {
-  const errors = validationResult(req)
-  if(!errors.isEmpty()){
-    return res.status(400).json({errors: errors.array()})
-  }
-  try {
-    const id = req.params.id
-    const { body } = req
-    const shop = await Shop.findOneAndUpdate({ _id: id , ownerId : req.user.id },body,{ new: true})
-    res.status(200).json(shop)
-  } catch (err) {
-    console.log(err)
-    res.status(500).json({ errors: "Internal server error" })
-  }
-}
+
 
 shopsCltr.getOneshop = async (req, res) => {
   try {
@@ -62,10 +48,43 @@ shopsCltr.getAllshop = async (req, res) => {
   }
 }
 
+shopsCltr.update = async (req, res) => {
+  const errors = validationResult(req)
+  if(!errors.isEmpty()){
+    return res.status(400).json({errors: errors.array()})
+  }
+  try {
+    const id = req.params.id
+    const body = _.pick(req.body,['ownerId','shopName','address','location','contact','description'])
+    const shop = await Shop.findOneAndUpdate({ _id: id , ownerId : req.user.id },body,{ new: true})
+    res.status(200).json(shop)
+  } catch (err) {
+    console.log(err)
+    res.status(500).json({ errors: "Internal server error" })
+  }
+}
+
+shopsCltr.updateStatus = async(req,res)=>{
+  const errors = validationResult(req)
+  if(!errors.isEmpty()){
+    return res.status(400).json({errors:errors.array()})
+  }
+  try{
+    const id = req.params.id
+    const body = _.pick(req.body,['approvalStatus'])
+    const shop = await Shop.findOneAndUpdate({_id:id},body,{new:true})
+    res.status(200).json(shop)
+  }
+  catch(err){
+    console.log(err)
+    res.status(500).json({ errors: "Internal server error" })
+  }
+}
+
 shopsCltr.destroy = async (req, res) => {
   try {
     const id = req.params.id
-    const shop = await Shop.findOneAndDelete({ _id: id })
+    const shop = await Shop.findOneAndDelete({ _id: id,ownerId:req.user.id })
     res.json(shop)
   } catch (error) {
     console.error(error)
